# ITS-API/ITSAPI/Models/Booking.cs
using Newtonsoft.Json;

namespace ITSAPI.Models
{
    public class Booking
    {
        [JsonProperty(PropertyName = "id")]
        public string? Id { get; set; }

        [JsonProperty(PropertyName = "projectId")]
        public string? ProjectId { get; set; }

        [JsonProperty(PropertyName = "resourceId")]
        public string? ResourceId { get; set; }

        [JsonProperty(PropertyName = "date")]
        public DateTime Date { get; set; }

        [JsonProperty(PropertyName = "hours")]
        public double Hours { get; set; }
    }
}