# ITS-API/ITSAPI/Models/Project.cs
using Newtonsoft.Json;

namespace ITSAPI.Models
{
    public class Project
    {
        [JsonProperty(PropertyName = "id")]
        public string? Id { get; set; }

        [JsonProperty(PropertyName = "name")]
        public string? Name { get; set; }

        [JsonProperty(PropertyName = "description")]
        public string? Description { get; set; }

        [JsonProperty(PropertyName = "customer")]
        public string? Customer { get; set; }

        [JsonProperty(PropertyName = "startDate")]
        public DateTime StartDate { get; set; }

        [JsonProperty(PropertyName = "createdBy")]
        public string? CreatedBy { get; set; }
    }
}